add: add HeroSection in LivingLab page
